pyproject.toml
[tool.poetry]
name = "speech_neuron"
version = "0.0.1"
description = ""
authors = ["C. Thomas Brittain <cthomasbrittain@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
soundfile = "^0.13.1"
fastapi = {extras = ["standard"], version = "^0.115.8"}
pyyaml = "^6.0.2"
sounddevice = "^0.5.1"
kokoro-onnx = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
